# lawless/src/main/resources/swagger-ui/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API SPEC FOR LAWLESS IT SECURITY HOMEWORK WEBAPP",
    "version": "v1",
    "title": "REST API SPEC FOR LAWLESS IT SECURITY HOMEWORK WEBAPP"
  },
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Caff"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login the user",
        "description": "",
        "operationId": "login",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/LoginReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/LoginResp"
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get the user",
        "description": "",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/UserResp"
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a user",
        "description": "",
        "operationId": "register",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RegisterReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/RegisterResp"
            }
          }
        }
      }
    },
    "/api/caff/commentAdd": {
      "post": {
        "tags": ["Caff"],
        "summary": "Comment caff ",
        "description": "",
        "operationId": "comment",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CommentAddCaffReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DetailsAllCaffResp"
            }
          }
        }
      }
    },
    "/api/caff/create": {
      "post": {
        "tags": ["Caff"],
        "summary": "Create and process caff ",
        "description": "",
        "operationId": "create",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "file",
            "in": "query",
            "required": true,
            "type": "ref"
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "price",
            "in": "query",
            "required": true,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/CreateCaffResp"
            }
          }
        }
      }
    },
    "/api/caff/delete": {
      "post": {
        "tags": ["Caff"],
        "summary": "Delete caff ",
        "description": "",
        "operationId": "delete",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/DeleteCaffReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DeleteCaffResp"
            }
          }
        }
      }
    },
    "/api/caff/details": {
      "get": {
        "tags": ["Caff"],
        "summary": "Details caff ",
        "description": "",
        "operationId": "details",
        "parameters": [
          {
            "name": "",
            "in": "query",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DetailsCaffResp"
            }
          }
        }
      }
    },
    "/api/caff/detailsAll": {
      "get": {
        "tags": ["Caff"],
        "summary": "Details all caff ",
        "description": "",
        "operationId": "detailsAll",
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DetailsAllCaffResp"
            }
          }
        }
      }
    },
    "/api/caff/update": {
      "post": {
        "tags": ["Caff"],
        "summary": "Update caff ",
        "description": "",
        "operationId": "update",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/UpdateCaffResp"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Comment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "timeStamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CommentAddCaffReq": {
      "type": "object",
      "properties": {
        "caffId": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "CreateCaffResp": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "DeleteCaffReq": {
      "type": "object",
      "properties": {
        "caffId": {
          "type": "string"
        }
      }
    },
    "DeleteCaffResp": {
      "type": "object",
      "properties": {
        "response": {
          "type": "string"
        }
      }
    },
    "DetailsAllCaffResp": {
      "type": "object",
      "properties": {
        "detailsAllCaffResponse": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DetailsCaffResp"
          }
        }
      }
    },
    "DetailsCaffResp": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "uploaded": {
          "type": "string",
          "format": "date-time"
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "previewPicture": {
          "type": "string",
          "format": "byte"
        },
        "userPersonalData": {
          "$ref": "#/definitions/UserPersonalData"
        }
      }
    },
    "LoginReq": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "LoginResp": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "userPersonalData": {
          "$ref": "#/definitions/UserPersonalData"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          }
        }
      }
    },
    "RegisterReq": {
      "type": "object",
      "properties": {
        "userPersonalData": {
          "$ref": "#/definitions/UserPersonalData"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "RegisterResp": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "Role": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "enum": ["ROLE_USER", "ROLE_ADMIN"]
        }
      }
    },
    "UpdateCaffResp": {
      "type": "object",
      "properties": {
        "response": {
          "type": "string"
        }
      }
    },
    "UserPersonalData": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    },
    "UserResp": {
      "type": "object",
      "properties": {
        "userPersonalData": {
          "$ref": "#/definitions/UserPersonalData"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          }
        }
      }
    }
  }
}